Checksum content graph (#75)
The contents of the content graphs now gets hashed in the record builder
to compute a checksum. This checksums are stored on the metadata graph.
Before computing the checksum the content graphs are canonicalized.
Also fixed a bug where the metadata graph would have the following
triple `:recId :hasContent :_b0` even though no content for the content
graph was provided. If no content is provided then the metadata graph
should not contain this triple. It is therefore possible to create a
record consisting of only one metadata graph, and no content graph. I
updated the documentation to reflect this. Also updated the
documentation for the hashing.