Added TripleStore.FindRecords (#82)
Added `.FindRecords()` to let people parse collections of records out of
a TripleStore.

In the research for this we looked into [blank
nodes](https://www.w3.org/TR/rdf11-concepts/#section-blank-nodes) and
found that they are not supposed to be portable identifiers. This means
that any relations to blank nodes cannot be maintained in collections.

Now content graphs are no longer "identified" by blank nodes, but are
instead identified by the record's ID suffixed with `#content`.