Update Record version (#92)
[User Story
171628](https://dev.azure.com/EquinorASA/Spine/_workitems/edit/171628):
Update record provenance

Previously, the Record version could be found in the .csproj file. After
we made changes to publish the nuget package on a new release, the
version is no longer in the .csproj file, which is where we got the
version to put in the record provenance before. We now use commit url
instead.