Upgrade to .NET 9 and upgrade packages (#101)
###
[AB#238635](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/238635)
[Pull requests practices for the SSI
team](https://github.com/equinor/ssi-infrastructure/blob/main/docs/pr_practices.md)

## Aim of the PR
The aim of this PR is to upgrade from .NET 8 to .NET 9 and upgrade all
packages.

## Implementation 
This was done by upgrading to .NET 9 in all the C# projects, and also
upgrading dotnet-version in our Github actions workflow files. All
packages that hade newer versions have also been upgraded, including
major upgrades.

## Other changes
There was both a `doc` and a `docs` folder. I moved the
`pull_request_template.md` to the same folder as the other documents and
renamed it to `docs`, so that we only have one folder.