Canonicalisation is no longer the default (#105)
###
[AB#269801](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/269801)
This fixes
[AB#269801](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/269801)
[Pull requests practices for the SSI
team](https://github.com/equinor/ssi-infrastructure/blob/main/docs/pr_practices.md)

## Aim of the PR
This PR changes the default functionality of the `RecordBuilder` and
`Immutable.Record.ToString()` away from canonicalisation. This because
there exists cases where too many blank nodes will overwhelm dotNetRdf's
`RdfCanonicaliser`.

## Implementation 
Added a flag to the constructor of `RecordBuilder`, with a default of
`false`. This flag determines whether the content is hashed or not.

Removed `ToNonCanonString` from `Immutable.Record` and replaced it with
`ToString`. Added `ToCanonString` which has the same functionality that
`ToString` used to have.

## Type of change
- [X] Bug fix 
- [ ] New feature 
- [X] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
Please describe the tests that you ran to verify that your changes works
as expected.

## Additional Changes
Removed usages of obsolete methods in `Immutable.Record`.