Optional force link on describes (#109)
###
[AB#277473](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/277473)

## Aim of the PR
Enforce connection on describes on metadata graph to the content graph. 
The object that is being described must be present on the content graph!

## Implementation 
- Added a default parameter `bool enforceDescribes` in the constructor
of the RecordBuilder. The default value of this parameter is false, if
it was set to true then the changes would be breaking and we would have
to update a lot of code, especially tests.
- If this param is set to true, and the objects being describes is not
mentioned as subjects in the content graph an exception is thrown.
- 
## Type of change
- [ ] Bug fix 
- [x] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

We should update the Records nuget to this version and enable the
enforceDescribes when we create records.

## How Has This Been Tested?
Two new integration tests in place; one negative test, and one positive.

## Additional Changes
Some removal of unnecessary usings.

---------

Co-authored-by: Markus Andre Pedersen <89122836+markus-ap@users.noreply.github.com>
Co-authored-by: Dag Hovland <dhovl@equinor.com>