Bug in RecordBuilder WithAdditionalMetadata (#118)
###
[AB#348735](https://dev.azure.com/EquinorASA/bb9bd8cb-74f7-4ffa-b0cb-60eff0a0be58/_workitems/edit/348735)

## Aim of the PR
Fixed a bug in the RecordBuilder. WithAdditonalMetadata would also copy
all content from the contentgraph to the metadatagraph. The reason was
concatenating the additional metadata added with the content triples,
instead of using the metadata triples. I also renamed "Triples" in the
Storage to "ContentTriples" to make the separation between the content
and metadata more clear.

## Type of change
- [x] Bug fix 
- [ ] New feature 
- [ ] Breaking change 
- [ ] This change requires a documentation update

## How Has This Been Tested?
Created unit test to check that the data is no longer copied, the test
also checks that the additional metadata is actually added to the
metadatagraph.